tests fixes

+++ src/main/resources/db/migration/V1_27__add_assessment_status_table.sql
@@ -18,7 +18,7 @@ create table resettlement_assessment
     id                    serial constraint resettlement_assessment_pkey primary key,
     prisoner_id           integer not null references prisoner (id),
     pathway_id            integer not null references pathway (id),
-    assessment_status_id  numeric not null references assessment_status (id),
+    assessment_status_id  integer not null references assessment_status (id),
     assessment_type       varchar(50) not null,
     assessment            jsonb,
     status_changed_to_status_id integer references status (id),
